// web/src/api/models/question.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { QuestionType } from './question-type';

/**
 * 
 * @export
 * @interface Question
 */
export interface Question {
    /**
     * 
     * @type {string}
     * @memberof Question
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Question
     */
    'question_set_id'?: string;
    /**
     * 
     * @type {QuestionType}
     * @memberof Question
     */
    'question_type'?: QuestionType;
    /**
     * 
     * @type {string}
     * @memberof Question
     */
    'question_text'?: string;
    /**
     * 
     * @type {string}
     * @memberof Question
     */
    'ground_truth'?: string;
    /**
     * 
     * @type {string}
     * @memberof Question
     */
    'gmt_created'?: string;
    /**
     * 
     * @type {string}
     * @memberof Question
     */
    'gmt_updated'?: string;
}